split helper functions into their own module - lightens server.js considerably!
